pyproject.toml
[build-system]
requires = [
  "setuptools",
  "setuptools-scm",
  "ctypesgen",
]
build-backend = "setuptools.build_meta"

[project]
name = "tandamaster"
description = "A music player specialized for playing music at milongas"
dynamic = ["version"]
authors = [
  { name = "Sašo Živanović", email = "saso.zivanovic@guest.arnes.si"},
]
dependencies = [
  "PyQt5",
  "PyGObject",
  "mutagen",
  "Unidecode",
  "bidict",
  "systemd-python",
  "pydantic",
]
license = "GPL-3.0-or-later"
license-files = [ "LICENCE" ]
keywords = [ "tango" ]
classifiers = [
  "Topic :: Multimedia :: Sound/Audio :: Players",
  "Development Status :: 4 - Beta",
  "Environment :: X11 Applications :: Qt",
  "Programming Language :: Python",
  "Intended Audience :: End Users/Desktop",
]
readme = "README.md"

[project.optional-dependencies]
debug = ["IPython"]

[project.urls]
Repository = "https://github.com/sasozivanovic/tandamaster"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]

[project.gui-scripts]
tandamaster = "tandamaster.tandamaster:main"

[project.entry-points."pipx.run"]
tandamaster = "tandamaster.tandamaster:main"
